pyproject.toml
[tool.poetry]
name = "age-prediction"
version = "0.1.0"
description = "Model for age prediction"
authors = ["matheus-tx <matheus.tx30@gmail.com>"]
license = "None, I think"
readme = "README.md"
packages = [{include = "age_prediction"}]

[tool.poetry.dependencies]
python = ">=3.10, <3.12"
numpy = "1.24.3"
pillow = "^10.0.0"
omegaconf = "^2.3.0"
pre-commit = "^3.3.3"
hydra-core = "^1.3.2"
ipykernel = "^6.24.0"
scikit-learn = "^1.3.0"
tensorflow-probability = "^0.20.1"
keras-vggface = "^0.6"
keras-applications = "^1.0.8"
pandas = "^2.0.3"
typeguard = "^4.0.0"
tensorflow = ">=2.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
